--- pyproject.toml
@@ -1,24 +1,0 @@
-[tool.poetry]
-name = "cards"
-version = "0.1.0"
-description = "Flash cards at the command line"
-authors = ["Dan Dean","Will Dean"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.9"
-typer = {extras = ["all"], version = "^0.9.0"}
-numpy = "^2.0.1"
-sqlmodel = "^0.0.8"
-
-[tool.poetry.scripts]
-card = "cards.__main__:main"
-
-[tool.poetry.group.dev.dependencies]
-black = "^23.7.0"
-pytest = "^7.4.0"
-pytest-cov = "^4.1.0"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
